+++ packages/trigger-sdk/src/job.ts
@@ -4,11 +4,11 @@ import {
   IntegrationConfig,
   InvokeOptions,
   JobMetadata,
-  LogLevel,
   Prettify,
   RunNotification,
   SuccessfulRunNotification,
 } from "@trigger.dev/core";
+import { LogLevel } from "@trigger.dev/core-backend";
 import { ConcurrencyLimit } from "./concurrencyLimit";
 import { IOWithIntegrations, TriggerIntegration } from "./integrations";
 import { runLocalStorage } from "./runLocalStorage";