+++ integrations/plain/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @trigger.dev/plain
 
+## 2.3.2
+
+### Patch Changes
+
+- 0c14e4cd: Fixed importing package subpath
+  - @trigger.dev/integration-kit@2.3.2
+  - @trigger.dev/sdk@2.3.2
+
+## 2.3.1
+
+### Patch Changes
+
+- Updated dependencies [f3efcc0c]
+  - @trigger.dev/sdk@2.3.1
+  - @trigger.dev/integration-kit@2.3.1
+
 ## 2.3.0
 
 ### Minor Changes