+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @trigger.dev/nestjs
 
+## 2.3.2
+
+### Patch Changes
+
+- @trigger.dev/sdk@2.3.2
+
+## 2.3.1
+
+### Patch Changes
+
+- Updated dependencies [f3efcc0c]
+  - @trigger.dev/sdk@2.3.1
+
 ## 2.3.0
 
 ### Minor Changes