+++ packages/testing/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @trigger.dev/testing
 
+## 2.3.2
+
+### Patch Changes
+
+- @trigger.dev/core@2.3.2
+- @trigger.dev/sdk@2.3.2
+
+## 2.3.1
+
+### Patch Changes
+
+- Updated dependencies [f3efcc0c]
+  - @trigger.dev/sdk@2.3.1
+  - @trigger.dev/core@2.3.1
+
 ## 2.3.0
 
 ### Minor Changes